+++ net/src/mbedtls/mbedtls.c
@@ -664,8 +664,8 @@ static int send_ssl(avs_net_abstract_socket_t *socket_,
     }
 
     if (bytes_sent < buffer_length) {
-        LOG(ERROR, "send failed (%zu/%zu): %d",
-            bytes_sent, buffer_length, result);
+        LOG(ERROR, "send failed (%lu/%lu): %d",
+            (unsigned long) bytes_sent, (unsigned long) buffer_length, result);
         update_send_or_recv_error_code(socket, result);
         return -1;
     }
